+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added 
 
+- Add IBaseCustomApiFakeMessageExecutor to better support both strongly-typed and late bound customa api when registering them in middleware
 - Add new ICustomApiFakeMessageExecutor and IGenericCustomApiFakeMessageExecutor interfaces to allow custom API execution with an optional plugin type as part of the Main Operation stage
  
 ### Changed